New text system iteration
Everything is related to the new system: Code, materials, datatables with styles, etc.

The new system appends desired symbols from a datatable row to a text, then parses it to turn it into an array of printed chars, a map of symbol coordinates and a set of style ending coordinates.
The output is then parsed by a widget, which will properly apply text effects, events, parameters, etc.
Currently it is possible to:
- set styles
- set off a pause timer
- change text speed
- call an event (via dispatcher)
Currently it is not possible, but intended:
- properly wrap text based on words rather than characters
- formatting symbols, specially for new lines, could also be used for variables

-test: Copied Gibran's dialogue boxes classes, and adapted the copies to support this system. It loads, but the text won't show up (it WILL be parsed/commanded to print, however), and switching dialogues will set off an infinite loop, somehow.