(Possibly) Final Typewritter Text Iteration / Localization Base
feat(Typewritter Text): created yet another iteration of the typewritter text workflow, this may be the final version of the proccess
and next iterations will simply add to this one. This version combines the previous NeoKG version with the markdown parsing of
Estellarium's version.

-added [BP_TypewritterTextBlock], [WBP_DialogueBoxTest] and [WBP_DialogueWindowTest], the new versions of the previous classes (the
names are temporary as the old classes have not been deleted yet, they will probably be deleted in the next commit)

-the TextBox will now obtain all of the required text and variables from a single Text variable, separating the raw text into separated
segments, and them separating these segments into raw text and markdown:
|-<Markdown>A text segment</>
|-<Markdown>Another text segment with different formatting and variables</>

-the desired syntax of the formatted text is almost the same as Estellarium's last version, with minor tweaks:
|-() contains the rich text style to be applied to a segment of text
|-%% contains the delay in seconds to wait before printing the segment
|-## contains the speed to print the segment, as time in seconds between each character
|-[] contains an event to be called by the textbox, the event will be transmitted to the GameMode trough the dispatcher
[GlobalTextEvent], from wich it can be listened to by any other class; you can also send a variable (or multiple) by appending it with
a comma
|-{} when in the text and not on the markdown, it will request a variable from the textbox owner, wich must be mannualy specified and
must implement the [BPI_TextVariableSupplier] interface

-a custom syntax highlighter for notepad++ is being worked on for the scriptwritters

-more functionalities will be added to the system soon, probably in the very next commit, like:
|-setting different ui styles to the dialogue system (or other ui elements) based on the speaking character
|-changing the rich text styles based on the speaking character
|-including a markdown key to break a line
|-making possible for two characters to be displayed on the screen at once

feat(Localization): the base config files for the localization system have been added, altough no localization functionality is actually
implemented yet. The native language has been set to Portuguese (because of dialogue) with English as a translation, this may be reversed
in the future

style(Thumbnail): added a little image of the protagonists as the project thumbnail, this is only a placeholder during development

-refactor(Folders): moved around some texture assets

style(Folders): color-coded the personal folders for ease of navigation and aesthetic